new webpack configuration
