+++ package.json
@@ -33,10 +33,12 @@
         "@babel/preset-react": "^7.6.3",
         "@types/c3": "^0.7.8",
         "@types/d3": "^7.4.0",
+        "@types/node": "^20.12.7",
         "@types/papaparse": "^5.3.11",
         "@types/react": "^18.0.26",
         "@types/react-dom": "^18.0.9",
         "@types/react-redux": "7.1.16",
+        "@types/webpack": "^5.28.5",
         "@typescript-eslint/eslint-plugin": "^5.45.0",
         "@typescript-eslint/parser": "^5.45.0",
         "alex": ">=11.0.0",
@@ -53,7 +55,8 @@
         "style-loader": "^1.1.3",
         "terser-webpack-plugin": "^5.3.6",
         "ts-loader": "^8.3.0",
-        "typescript": "^4.8.4",
+        "ts-node": "^10.9.2",
+        "typescript": "^4.9.5",
         "webpack": "^5.76.0",
         "webpack-cli": "^4.10.0"
     },
